feat(310p)/docs: P3 eager E2E acceptance - Qwen3.6 TP=2 runs with fused o_proj
Real-machine results on the dual-die 310P3 (FP16, spawn workers):

- Model loads (18.85 GB weights, 7.9 GiB KV cache) and warmup passes with
  zero errors; the full-attention o_proj layers are confirmed on the fused
  path (worker log 'Enable 310P3 TP=2 MemFabric unquantized full-attention
  o_proj pipeline (tile_m=32, chunk_bytes=131072, FP16 exchange)') while
  linear-attention layers stay on the stock path by construction.
- 3 rounds x 4 prompts (12 generates) all succeed; same-prompt outputs are
  identical across rounds (end-to-end repeated-wave stability).
- Feature-off baseline comparison: 10/12 outputs are byte-identical; the 2
  remaining differ at single-token granularity (e.g. one conjunction word),
  traced to HCCL allreduce's internal accumulation precision vs the fused
  FP32-math reduce - expected numeric behavior, not a correctness bug (the
  single-layer acceptance already proved bit-exactness vs FP32-math
  semantics).
- bench_310p_memfabric_o_proj.py communication benchmark moved to the FP16
  payload contract and re-verified (rows 1..2048 PASS).
- Fixes along the way: log wording still said 'W8A8'; /dev/shm raised to 2G
  for the vllm shm broadcast; VLLM_WORKER_MULTIPROC_METHOD=spawn required
  because fork workers make the MemFabric-linked dlopen relocation crawl
  (10min+); numba installed for the ngram proposer import.
- dev-plan 4.12 records the round, including the upstream wgm-dev-310p
  auto-deploy (KFC channel) integration facts.

+++ benchmarks/scripts/bench_310p_memfabric_o_proj.py
@@ -35,12 +35,12 @@ def _sync() -> None:
 
 
 def _make_local(rows: int, rank: int, device: torch.device) -> torch.Tensor:
-    # Values are exactly representable enough for a deterministic BF16 sum while
-    # still varying by row/rank. Keep magnitude small to avoid reduction noise.
+    # Values are exactly representable in FP16 while still varying by row/rank.
+    # Keep magnitude small to avoid reduction noise.
     row = torch.arange(rows, device=device, dtype=torch.float32).unsqueeze(1)
     col = torch.arange(WIDTH, device=device, dtype=torch.float32).unsqueeze(0)
     value = (rank + 1) * 0.25 + (row % 31) * 0.001 + (col % 17) * 0.0001
-    return value.to(torch.bfloat16)
+    return value.to(torch.float16)
 
 
 def _runtime_ops():
@@ -100,12 +100,11 @@ def _time_ms(fn, warmup: int, repeat: int) -> tuple[float, float, float]:
 
 
 def _hccl_once(local: torch.Tensor) -> torch.Tensor:
-    # 310P HCCL has no BF16 allreduce (CANN 9.1: HcclAllreduce rejects BF16).
-    # The reference reduces in FP32 and rounds back to BF16, which matches the
-    # MemFabric reduce kernel semantics (BF16 -> F32 add -> BF16 round).
+    # The reference reduces in FP32 and rounds back to FP16, which matches the
+    # MemFabric reduce kernel semantics (FP16 -> F32 add -> FP16 round).
     out = local.float()
     dist.all_reduce(out)
-    return out.to(torch.bfloat16)
+    return out.to(torch.float16)
 
 
 def main() -> None:
